// types/form.ts
export type Form714Data = {
  respondent_identification_code: string
  respondent_name: string
  mailing_address: string
  contact_name: string
  contact_title: string
  contact_email_address: string
  contact_telephone_number: string
  certifying_official_name: string
  certifying_official_title: string
  certifying_official_date: string
  name: string
}

export type Form6QData = {
  respondent_exact_legal_name: string
  year_period_report: string
  previous_name_and_change_date: string
  principal_office_address: string
  contact_person_name: string
  contact_person_title: string
  contact_person_address: string
  contact_person_phone: string
  date_of_report: string
}